# services/starter-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST APIs for managing starters and organizations.",
        "title": "Starter Service API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/starters": {
            "get": {
                "description": "Retrieve starters with optional search and pagination",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Starters"
                ],
                "summary": "List starters",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Keyword search",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "fullname",
                            "domain",
                            "dept_name",
                            "bu_name"
                        ],
                        "type": "string",
                        "description": "Search field",
                        "name": "search_by",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "id",
                            "domain",
                            "created_at"
                        ],
                        "type": "string",
                        "default": "id",
                        "description": "Sort field",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "default": "asc",
                        "description": "Sort order",
                        "name": "sort_order",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 20,
                        "description": "Page size",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/starter.StarterListAPIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/shared.GenericAPIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/shared.GenericAPIResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new starter record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Starters"
                ],
                "summary": "Create starter",
                "parameters": [
                    {
                        "description": "Starter payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/starter.CreateStarterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/starter.StarterAPIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/shared.GenericAPIResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/shared.GenericAPIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/shared.GenericAPIResponse"
                        }
                    }
                }
            }
        },
        "/starters/{domain}": {
            "get": {
                "description": "Retrieve a starter by domain",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Starters"
                ],
                "summary": "Get starter detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Starter domain",
                        "name": "domain",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/starter.StarterAPIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/shared.GenericAPIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/shared.GenericAPIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/shared.GenericAPIResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Soft delete a starter by domain",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Starters"
                ],
                "summary": "Delete starter",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Starter domain",
                        "name": "domain",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/starter.StarterDeleteAPIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/shared.GenericAPIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/shared.GenericAPIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/shared.GenericAPIResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Partially update a starter by domain",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Starters"
                ],
                "summary": "Update starter",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Starter domain",
                        "name": "domain",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/starter.UpdateStarterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/starter.StarterAPIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/shared.GenericAPIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/shared.GenericAPIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/shared.GenericAPIResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "httputil.RespPagination": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "next": {
                    "type": "string"
                },
                "prev": {
                    "type": "string"
                },
                "total_items": {
                    "type": "integer"
                }
            }
        },
        "shared.BusinessUnitNested": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "shortname": {
                    "type": "string"
                }
            }
        },
...
        "shared.DepartmentNested": {
            "type": "object",
            "properties": {
                "group_department": {
                    "$ref": "#/definitions/shared.GroupDepartmentNested"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "shortname": {
                    "type": "string"
                }
            }
        },
        "shared.GenericAPIResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "error": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "shared.GroupDepartmentNested": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "shortname": {
                    "type": "string"
                }
            }
        },
        "shared.LineManagerNested": {
            "type": "object",
            "properties": {
                "domain": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "job_title": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "starter.CreateStarterRequest": {
            "type": "object",
            "required": [
                "department_id",
                "domain",
                "email",
                "job_title",
                "mobile",
                "name",
                "work_phone"
            ],
            "properties": {
                "department_id": {
                    "type": "integer"
                },
                "domain": {
                    "type": "string",
                    "maxLength": 25,
                    "minLength": 3
                },
                "email": {
                    "type": "string",
                    "maxLength": 100
                },
                "job_title": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                },
                "line_manager_id": {
                    "type": "integer"
                },
                "mobile": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 10
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 2
                },
                "work_phone": {
                    "type": "string",
                    "maxLength": 20
                }
            }
        },
        "starter.StarterAPIResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/starter.StarterResponse"
                },
                "error": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "starter.StarterDeleteAPIResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/starter.StarterDeleteResponseData"
                },
                "error": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "starter.StarterDeleteResponseData": {
            "type": "object",
            "properties": {
                "domain": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "starter.StarterListAPIResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/starter.StarterListAPIResponseData"
                },
                "error": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "starter.StarterListAPIResponseData": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/starter.StarterResponse"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/httputil.RespPagination"
                }
            }
        },
        "starter.StarterResponse": {
            "type": "object",
            "properties": {
                "business_unit": {
                    "$ref": "#/definitions/shared.BusinessUnitNested"
                },
                "created_at": {
                    "type": "string"
                },
                "department": {
                    "$ref": "#/definitions/shared.DepartmentNested"
                },
                "domain": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "job_title": {
                    "type": "string"
                },
                "line_manager": {
                    "$ref": "#/definitions/shared.LineManagerNested"
                },
                "mobile": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "work_phone": {
                    "type": "string"
                }
            }
        },
        "starter.UpdateStarterRequest": {
            "type": "object",
            "required": [
                "domain"
            ],
            "properties": {
                "department_id": {
                    "type": "integer"
                },
                "domain": {
                    "type": "string",
                    "maxLength": 25,
                    "minLength": 2
                },
                "email": {
                    "type": "string",
                    "maxLength": 100
                },
                "job_title": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                },
                "line_manager_id": {
                    "type": "integer"
                },
                "mobile": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 10
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 2
                },
                "work_phone": {
                    "type": "string",
                    "maxLength": 20
                }
            }
        }
    }
}