# services/starter-service/internal/starter/presentation/http/dto/shared/api_response.go
package shared

// GenericAPIResponse describes the common envelope returned by the API.
type GenericAPIResponse struct {
	Code    int         `json:"code"`
	Message string      `json:"message"`
	Data    interface{} `json:"data,omitempty"`
	Error   interface{} `json:"error,omitempty"`
}